requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlmodel==0.0.24
psycopg2==2.9.10
pydantic[email]
python-multipart==0.0.20
PyJWT
passlib[bcrypt]
python-dotenv==1.1.1